more fixes plust started work on nifas

+++ src/test/kotlin/LogicTest.kt
@@ -33,9 +33,9 @@ class LogicTest {
     @Test
     fun testRemoveTuhrLessThan15(){
         val fixedDurations = mutableListOf(
-            FixedDuration(DurationType.DAM, timeInMilliseconds = (MILLISECONDS_IN_A_DAY*2).toLong()),
-            FixedDuration(DurationType.TUHR, timeInMilliseconds = (MILLISECONDS_IN_A_DAY*2).toLong()),
-            FixedDuration(DurationType.DAM, timeInMilliseconds = (MILLISECONDS_IN_A_DAY*2).toLong())
+            FixedDuration(DurationType.DAM, timeInMilliseconds = (MILLISECONDS_IN_A_DAY*2)),
+            FixedDuration(DurationType.TUHR, timeInMilliseconds = (MILLISECONDS_IN_A_DAY*2)),
+            FixedDuration(DurationType.DAM, timeInMilliseconds = (MILLISECONDS_IN_A_DAY*2))
         )
         removeTuhrLessThan15(fixedDurations)
         assertEquals(1, fixedDurations.size)
